Move oauth and callback endpoint to oauth.py
